+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 ### Fixed
 
 - Improve functional path checking.
+- Fix stats refresh issue in popup.
 
 ### Changed
 
 - Change keyboard shorcut visual.
--
+- Improve issue tracking.
+
 
 ## [4.0.1] - 2023-05-04